move mutable indexes into dedicated package
