+++ pinot-segment-local/src/main/java/org/apache/pinot/segment/local/realtime/impl/forward/VarByteSVMutableForwardIndex.java
@@ -20,7 +20,7 @@ package org.apache.pinot.segment.local.realtime.impl.forward;
 
 import java.io.IOException;
 import org.apache.pinot.segment.local.io.writer.impl.MutableOffHeapByteArrayStore;
-import org.apache.pinot.segment.spi.index.reader.MutableForwardIndex;
+import org.apache.pinot.segment.spi.index.mutable.MutableForwardIndex;
 import org.apache.pinot.segment.spi.memory.PinotDataBufferMemoryManager;
 import org.apache.pinot.spi.data.FieldSpec.DataType;
