+++ pinot-segment-local/src/main/java/org/apache/pinot/segment/local/realtime/impl/nullvalue/MutableNullValueVector.java
@@ -18,7 +18,7 @@
  */
 package org.apache.pinot.segment.local.realtime.impl.nullvalue;
 
-import org.apache.pinot.segment.spi.index.ThreadSafeMutableRoaringBitmap;
+import org.apache.pinot.segment.spi.index.mutable.ThreadSafeMutableRoaringBitmap;
 import org.apache.pinot.segment.spi.index.reader.NullValueVectorReader;
 import org.roaringbitmap.buffer.ImmutableRoaringBitmap;
