+++ pinot-segment-local/src/main/java/org/apache/pinot/segment/local/segment/index/readers/ValidDocIndexReaderImpl.java
@@ -18,7 +18,7 @@
  */
 package org.apache.pinot.segment.local.segment.index.readers;
 
-import org.apache.pinot.segment.spi.index.ThreadSafeMutableRoaringBitmap;
+import org.apache.pinot.segment.spi.index.mutable.ThreadSafeMutableRoaringBitmap;
 import org.apache.pinot.segment.spi.index.reader.ValidDocIndexReader;
 import org.roaringbitmap.buffer.ImmutableRoaringBitmap;
