+++ pinot-segment-spi/src/main/java/org/apache/pinot/segment/spi/index/mutable/ThreadSafeMutableRoaringBitmap.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.segment.spi.index;
+package org.apache.pinot.segment.spi.index.mutable;
 
 import org.roaringbitmap.buffer.MutableRoaringBitmap;
